+++ src/main/java/com/cloudhopper/smpp/channel/SmppChannelConstants.java
@@ -38,5 +38,6 @@ public class SmppChannelConstants {
     public static final String PIPELINE_SESSION_PDU_DECODER_NAME = "smppSessionPduDecoder";
     public static final String PIPELINE_SESSION_WRAPPER_NAME = "smppSessionWrapper";
     public static final String PIPELINE_SESSION_SSL_NAME = "smppSessionSSL"; 
+    public static final String PIPELINE_SESSION_WRITE_TIMEOUT_NAME = "smppSessionWriteTimeout";
 
 }